clear warnnings

+++ index/impl/gamma_index_binary_ivf.cc
@@ -209,19 +209,19 @@ int GammaIndexBinaryIVF::Indexing() {
     return 0;
   }
   RawVector *raw_vec = dynamic_cast<RawVector *>(vector_);
-  int vectors_count = raw_vec->MetaInfo()->Size();
+  size_t vectors_count = raw_vec->MetaInfo()->Size();
 
   size_t num;
-  if (indexing_size_ < nlist) {
+  if ((size_t)indexing_size_ < nlist) {
     num = nlist * 39;
     LOG(WARNING) << "Because index_size[" << indexing_size_ << "] < ncentroids[" << nlist 
                  << "], index_size becomes ncentroids * 39[" << num << "].";
-  } else if (indexing_size_ <= nlist * 265) {
-    if (indexing_size_ < nlist * 39) {
+  } else if ((size_t)indexing_size_ <= nlist * 265) {
+    if ((size_t)indexing_size_ < nlist * 39) {
       LOG(WARNING) << "Index_size[" << indexing_size_ << "] is too small. "
                    << "The appropriate range is [ncentroids * 39, ncentroids * 256]"; 
     }
-    num = indexing_size_;
+    num = (size_t)indexing_size_;
   } else {
     num = nlist * 256;
     LOG(WARNING) << "Index_size[" << indexing_size_ << "] is too big. "
