+++ index/impl/gamma_index_ivfpq.cc
@@ -473,16 +473,16 @@ int GammaIVFPQIndex::Indexing() {
   size_t vectors_count = raw_vec->MetaInfo()->Size();
 
   size_t num;
-  if (indexing_size_ < nlist) {
+  if ((size_t)indexing_size_ < nlist) {
     num = nlist * 39;
     LOG(WARNING) << "Because index_size[" << indexing_size_ << "] < ncentroids[" << nlist 
                  << "], index_size becomes ncentroids * 39[" << num << "].";
-  } else if (indexing_size_ <= nlist * 265) {
-    if (indexing_size_ < nlist * 39) {
+  } else if ((size_t)indexing_size_ <= nlist * 265) {
+    if ((size_t)indexing_size_ < nlist * 39) {
       LOG(WARNING) << "Index_size[" << indexing_size_ << "] is too small. "
                    << "The appropriate range is [ncentroids * 39, ncentroids * 256]"; 
     }
-    num = indexing_size_;
+    num = (size_t)indexing_size_;
   } else {
     num = nlist * 256;
     LOG(WARNING) << "Index_size[" << indexing_size_ << "] is too big. "
@@ -1363,7 +1363,7 @@ int GammaIVFPQIndex::Load(const std::string &index_dir) {
   } else if (ret == 0) {
     READ1(indexed_vec_count_);
     if (indexed_vec_count_ < 0 ||
-        indexed_vec_count_ > vector_->MetaInfo()->size_) {
+        indexed_vec_count_ > (int)vector_->MetaInfo()->size_) {
       LOG(ERROR) << "invalid indexed count [" << indexed_vec_count_ 
                  << "] vector size [" << vector_->MetaInfo()->size_ << "]";
       return INTERNAL_ERR;
